rename: 박스스코어를 boxScore -> boxscore로 통일(#111)
