+++ src/features/game/index.ts
@@ -1,4 +1,4 @@
-export { BoxScoreTab } from './BoxScoreTab';
+export { BoxscoreTab } from './BoxscoreTab';
 export { MatchScheduleTab } from './MatchScheduleTab';
 export { WatchPointTab } from './WatchPointTab';
 export { RankingTab } from './RankingTab';
