+++ src/pages/game/regular/RegularGamePage.tsx
@@ -5,7 +5,7 @@ import { cn } from '@/lib/utils';
 
 import Layout from '@/features/common/Layout';
 import {
-  BoxScoreTab,
+  BoxscoreTab,
   MatchScheduleTab,
   RankingTab,
   WatchPointTab,
@@ -94,7 +94,7 @@ function RegularGamePage() {
           <MatchScheduleTab />
         </TabsContent>
         <TabsContent value="boxscore">
-          <BoxScoreTab />
+          <BoxscoreTab />
         </TabsContent>
         <TabsContent value="ranking">
           <RankingTab />
